rewrite frameskipping code
Instead of measuring time, averaging it and trying to adjust to meet the
timing goals next frame, we simply frameskip until the timing goal is
reached. To prevent too extreme movements in frameskip, we skip at least
one less than in the previous frame, and at most two more.
This is far from perfect when sound is enabled (there might be a slow
interrupt handler involved), so we may have to do some averaging despite
having a high-precision timer now.