use 4Ntv and emuIf input, do away with "native" input
NativeGE_getKeyInput() does not work for serially-attached controllers
and is therefore not available at all on some systems. Use 4Ntv instead
to get a minimal set of known keys and emuIfKeyGetInput() for the rest.
(The latter interface provides a key mapping, but we haven't figured
it out yet, so we just use the raw keys for now.)

The horrible runtime behavior of NativeGE_getKeyInput4Ntv() on the A1000
is likely the result of debug output being written to the serial port.
D'oh! Turning debug output off globally should fix it.